requirements.txt
chardet==3.0.4
coverage==5.3
idna==2.10
Pillow==8.0.1
pony==0.7.14
psycopg2==2.8.6
requests==2.25.0
six==1.15.0
urllib3==1.26.2
vk-api==11.9.0

